chore(Comscore): move sdk loader to new file
